refactor(test-harness): one type per file — split the verdict layer
verdict.rs held six types. Each moves to its own module under output/summary/:
OutputVerdict, SizeChange, GasChange, FailureVerdict, SuiteFailures, HealthIssue.
The from_stats constructors move with their type; the line() rendering impls
stay in render.rs for now. The shared test constructors become methods on the
types they build — SuiteStats::available and DiffCounter::counted — rather than
free functions.
